chore: immutable llama call
